Moved danger utils. Moved smart_cover utils.
