+++ src/engine/core/schemes/stalker/combat_ignore/CombatProcessEnemyManager.ts
@@ -2,8 +2,8 @@ import { registry } from "@/engine/core/database";
 import { AbstractSchemeManager } from "@/engine/core/objects/ai/scheme";
 import { SmartTerrain } from "@/engine/core/objects/server/smart_terrain/SmartTerrain";
 import { ISchemeCombatIgnoreState } from "@/engine/core/schemes/stalker/combat_ignore/index";
+import { canObjectSelectAsEnemy } from "@/engine/core/schemes/stalker/danger/utils";
 import { LuaLogger } from "@/engine/core/utils/logging";
-import { canObjectSelectAsEnemy } from "@/engine/core/utils/object";
 import { startSmartTerrainAlarm } from "@/engine/core/utils/smart_terrain";
 import { logicsConfig } from "@/engine/lib/configs/LogicsConfig";
 import { ACTOR_ID } from "@/engine/lib/constants/ids";
