+++ src/engine/core/schemes/stalker/danger/DangerManager.ts
@@ -2,8 +2,8 @@ import { level, snd_type, time_global } from "xray16";
 
 import { AbstractSchemeManager } from "@/engine/core/objects/ai/scheme";
 import { ISchemeDangerState } from "@/engine/core/schemes/stalker/danger/ISchemeDangerState";
+import { canObjectSelectAsEnemy } from "@/engine/core/schemes/stalker/danger/utils";
 import { LuaLogger } from "@/engine/core/utils/logging";
-import { canObjectSelectAsEnemy } from "@/engine/core/utils/object";
 import { isSoundType } from "@/engine/core/utils/sound";
 import { logicsConfig } from "@/engine/lib/configs/LogicsConfig";
 import {
