+++ src/engine/core/schemes/stalker/danger/evaluators/EvaluatorDanger.ts
@@ -4,7 +4,7 @@ import { registry } from "@/engine/core/database";
 import { EActionId } from "@/engine/core/objects/ai/types";
 import { SmartTerrain } from "@/engine/core/objects/server/smart_terrain";
 import { ISchemeDangerState } from "@/engine/core/schemes/stalker/danger";
-import { isObjectFacingDanger } from "@/engine/core/utils/object";
+import { isObjectFacingDanger } from "@/engine/core/schemes/stalker/danger/utils";
 import { startSmartTerrainAlarm } from "@/engine/core/utils/smart_terrain";
 import { logicsConfig } from "@/engine/lib/configs/LogicsConfig";
 import { MAX_U16 } from "@/engine/lib/constants/memory";
