+++ src/engine/core/schemes/stalker/danger/utils/danger_generic.ts
