+++ src/engine/core/utils/object/index.ts
@@ -1,13 +1,11 @@
 export * from "@/engine/core/utils/object/object_action";
 export * from "@/engine/core/utils/object/object_check";
-export * from "@/engine/core/utils/object/object_danger";
 export * from "@/engine/core/utils/object/object_find";
 export * from "@/engine/core/utils/object/object_get";
 export * from "@/engine/core/utils/object/object_loot";
 export * from "@/engine/core/utils/object/object_meeting";
 export * from "@/engine/core/utils/object/object_set";
 export * from "@/engine/core/utils/object/object_setup";
-export * from "@/engine/core/utils/object/object_smart_cover";
 export * from "@/engine/core/utils/object/object_sound";
 export * from "@/engine/core/utils/object/object_state";
 export * from "@/engine/core/utils/object/object_weapon";
