+++ src/engine/core/utils/smart_cover.test.ts
