+++ src/engine/core/utils/smart_cover.ts
